+++ shake/compiler/lexer/src/commonMain/kotlin/io/github/shakelang/shake/lexer/token/stream/ShakeTokenInputStream.kt
@@ -1,6 +1,6 @@
 package io.github.shakelang.shake.lexer.token.stream
 
-import io.github.shakelang.parseutils.lexer.token.stream.TokenInputStream
+import io.github.shakelang.shake.util.parseutils.lexer.token.stream.TokenInputStream
 import io.github.shakelang.shake.lexer.token.ShakeToken
 import io.github.shakelang.shake.lexer.token.ShakeTokenType