+++ shake/compiler/lexer/src/commonTest/kotlin/io/github/shakelang/shake/lexer/token/stream/TokenBasedTokenInputStreamTests.kt
@@ -1,7 +1,7 @@
 package io.github.shakelang.shake.lexer.token.stream
 
-import io.github.shakelang.parseutils.characters.position.PositionMap
-import io.github.shakelang.parseutils.characters.source.CharacterSource
+import io.github.shakelang.shake.util.parseutils.characters.position.PositionMap
+import io.github.shakelang.shake.util.parseutils.characters.source.CharacterSource
 import io.github.shakelang.shake.lexer.token.ShakeToken
 import io.github.shakelang.shake.lexer.token.ShakeTokenType
 import kotlin.test.Test