+++ shake/compiler/parser/src/commonMain/kotlin/io/github/shakelang/shake/parser/node/ShakeIfNode.kt
@@ -1,6 +1,6 @@
 package io.github.shakelang.shake.parser.node
 
-import io.github.shakelang.parseutils.characters.position.PositionMap
+import io.github.shakelang.shake.util.parseutils.characters.position.PositionMap
 
 class ShakeIfNode(map: PositionMap, val body: ShakeBlockNode, val elseBody: ShakeBlockNode?, val condition: ShakeValuedNode) : ShakeValuedStatementNodeImpl(map) {