+++ shake/compiler/parser/src/commonMain/kotlin/io/github/shakelang/shake/parser/node/ShakeNode.kt
@@ -1,7 +1,7 @@
 package io.github.shakelang.shake.parser.node
 
-import io.github.shakelang.parseutils.characters.position.PositionMap
-import io.github.shakelang.shason.JSON
+import io.github.shakelang.shake.util.parseutils.characters.position.PositionMap
+import io.github.shakelang.shake.util.shason.JSON
 
 interface ShakeNode {
 
@@ -11,7 +11,7 @@ interface ShakeNode {
 }
 
 abstract class ShakeNodeImpl protected constructor(val map: PositionMap) : ShakeNode {
-    override fun toString(): String = JSON.stringify(this.json)
+    override fun toString(): String = io.github.shakelang.shake.util.shason.JSON.stringify(this.json)
 }
 
 interface ShakeStatementNode : ShakeNode