+++ shake/compiler/parser/src/commonMain/kotlin/io/github/shakelang/shake/parser/node/factor/ShakeDoubleNode.kt
@@ -1,6 +1,6 @@
 package io.github.shakelang.shake.parser.node.factor
 
-import io.github.shakelang.parseutils.characters.position.PositionMap
+import io.github.shakelang.shake.util.parseutils.characters.position.PositionMap
 import io.github.shakelang.shake.parser.node.ShakeValuedNodeImpl
 
 class ShakeDoubleNode(map: PositionMap, val number: Double) : ShakeValuedNodeImpl(map) {