+++ shake/shasambly/shasambly/src/commonMain/kotlin/io/github/shakelang/shake/shasambly/interpreter/ShasamblyOpcodeExecutor.kt
@@ -1,8 +1,9 @@
 @file:Suppress("nothing_to_inline")
 package io.github.shakelang.shake.shasambly.interpreter
 
-import io.github.shakelang.primitives.bytes.*
+import io.github.shakelang.shake.util.primitives.bytes.*
 import io.github.shakelang.shake.shasambly.interpreter.natives.nativeFunctions
+import io.github.shakelang.shake.util.primitives.bytes.*
 import kotlin.experimental.and
 import kotlin.experimental.or
 import kotlin.experimental.xor