+++ shake/shasambly/shasambly/src/commonTest/kotlin/io/github/shakelang/shake/shasambly/interpreter/ShasamblyInterpretingBaseStackTests.kt
@@ -1,6 +1,6 @@
 package io.github.shakelang.shake.shasambly.interpreter
 
-import io.github.shakelang.primitives.bytes.toBytes
+import io.github.shakelang.shake.util.primitives.bytes.toBytes
 import io.github.shakelang.shake.shasambly.interpreter.natives.Natives
 import kotlin.test.Test
 import kotlin.test.assertEquals