+++ shake/shasambly/shasp/src/commonMain/kotlin/io/github/shakelang/shake/shasambly/shasp/lexer/ShasPLexer.kt
@@ -1,6 +1,6 @@
 package io.github.shakelang.shake.shasambly.shasp.lexer
 
-import io.github.shakelang.parseutils.characters.streaming.CharacterInputStream
+import io.github.shakelang.shake.util.parseutils.characters.streaming.CharacterInputStream
 import io.github.shakelang.shake.lexer.token.ShasPToken
 import io.github.shakelang.shake.shasambly.shasp.lexer.token.OnDemandLexingShasPTokenInputStream
 import io.github.shakelang.shake.shasambly.shasp.lexer.token.ShasPTokenBasedInputStream