+++ shake/shasambly/shastools/src/commonMain/kotlin/io/github/shakelang/shake/shasambly/generator/basic/ShasamblyGenerator.kt
@@ -1,8 +1,8 @@
 @file: Suppress("unused")
 package io.github.shakelang.shake.shasambly.generator.basic
 
-import io.github.shakelang.primitives.bytes.toBytes
-import io.github.shakelang.primitives.bytes.toHexString
+import io.github.shakelang.shake.util.primitives.bytes.toBytes
+import io.github.shakelang.shake.util.primitives.bytes.toHexString
 import io.github.shakelang.shake.shasambly.interpreter.Opcodes
 
 /**