+++ shake/shasambly/shastools/src/commonMain/kotlin/io/github/shakelang/shake/shasambly/generator/simple/SimpleShasamblyGenerator.kt
@@ -1,7 +1,7 @@
 @file:Suppress("unused")
 package io.github.shakelang.shake.shasambly.generator.simple
 
-import io.github.shakelang.primitives.bytes.toBytes
+import io.github.shakelang.shake.util.primitives.bytes.toBytes
 import io.github.shakelang.shake.shasambly.generator.basic.*
 import io.github.shakelang.shake.shasambly.interpreter.natives.Natives