+++ util/logger/src/jsMain/kotlin/io/github/shakelang/shake/util/logger/Logger.kt
@@ -1,0 +1,76 @@
+package io.github.shakelang.shake.util.logger
+
+import io.github.shakelang.shake.util.colorlib.functional.blue
+import io.github.shakelang.shake.util.colorlib.functional.cyan
+import io.github.shakelang.shake.util.colorlib.functional.red
+import io.github.shakelang.shake.util.colorlib.functional.yellow
+import io.github.shakelang.shake.util.environment.getRunningEnvironment
+import kotlin.js.Console
+
+inline fun Console.debug(vararg message: String): Unit = asDynamic().debug(message) as Unit
+
+object JsColoredBrowserConsoleLogger : LoggerPipe() {
+
+    override fun log(level: LogLevel, message: String) {
+        when (level) {
+            LogLevel.INFO -> printInfo(message)
+            LogLevel.DEBUG -> printDebug(message)
+            LogLevel.SUCCESS -> printSuccess(message)
+            LogLevel.WARN -> printWarn(message)
+            LogLevel.ERROR -> printError(message)
+            LogLevel.FATAL -> printFatal(message)
+        }
+    }
+
+    private fun printInfo(message: String) = console.info("%c$message", "color: cyan;")
+    private fun printDebug(message: String) = console.debug("%c$message", "color: blue;")
+    private fun printSuccess(message: String) = console.debug("%c$message", "color: green;")
+    private fun printWarn(message: String) = console.warn("%c$message", "color: yellow;")
+    private fun printError(message: String) = console.error("%c$message", "color: red;")
+    private fun printFatal(message: String) = console.error("%c$message", "color: red;")
+
+}
+
+object JsColoredConsoleLogger : LoggerPipe() {
+
+    override fun log(level: LogLevel, message: String) {
+        when (level) {
+            LogLevel.INFO -> printInfo(message)
+            LogLevel.DEBUG -> printDebug(message)
+            LogLevel.SUCCESS -> printSuccess(message)
+            LogLevel.WARN -> printWarn(message)
+            LogLevel.ERROR -> printError(message)
+            LogLevel.FATAL -> printFatal(message)
+        }
+    }
+
+    private fun printInfo(message: String) = console.info(message.cyan().string())
+    private fun printDebug(message: String) = console.debug(message.blue().string())
+    private fun printSuccess(message: String) = console.debug(message.blue().string())
+    private fun printWarn(message: String) = console.warn(message.yellow().string())
+    private fun printError(message: String) = console.error(message.red().string())
+    private fun printFatal(message: String) = console.error(message.red().string())
+
+}
+
+object JsConsoleLogger: LoggerPipe() {
+
+    override fun log(level: LogLevel, message: String) {
+        when (level) {
+            LogLevel.INFO -> printInfo(message)
+            LogLevel.DEBUG -> printDebug(message)
+            LogLevel.SUCCESS -> printSuccess(message)
+            LogLevel.WARN -> printWarn(message)
+            LogLevel.ERROR -> printError(message)
+            LogLevel.FATAL -> printFatal(message)
+        }
+    }
+
+    private fun printInfo(message: String) = console.info(message)
+    private fun printDebug(message: String) = console.debug(message)
+    private fun printSuccess(message: String) = console.debug(message)
+    private fun printWarn(message: String) = console.warn(message)
+    private fun printError(message: String) = console.error(message)
+    private fun printFatal(message: String) = console.error(message)
+
+}