+++ util/common-io/changelog.md
@@ -1,5 +1,10 @@
 # Changelog for `common-io`
 
+## 0.1.2 (2023-11-01)
+
+- Move `io.github.shakelang.io` to `io.github.shakelang.shake.util.io`
+
+
 ## 0.1.1
 
 - Added `CountingOutputStream#byteCount` and `CountingInputStream#operationCount`